refactor: replace Cube component with dynamic import for ShatterCube; optimize for server-side rendering
